chore: removed duplicate session code and tidy up comments
